requirements.txt
agno
openai
duckduckgo-search
yfinance
sqlalchemy
fastapi
uvicorn
exa_py
lancedb
tantivy
pypdf
duckduckgo-search
googlesearch-python
pycountry
youtube_search